requirements.txt
streamlit
google-generativeai
Pillow
pypdf
docx2txt